Island: Move AgentPluginRepository into agent_plugin_service
