+++ monkey/monkey_island/cc/repositories/__init__.py
@@ -9,15 +9,11 @@ from .i_machine_repository import IMachineRepository
 from .i_agent_repository import IAgentRepository
 from .i_node_repository import INodeRepository
 from .i_agent_event_repository import IAgentEventRepository
-from .i_agent_plugin_repository import IAgentPluginRepository
 
 from .local_storage_file_repository import LocalStorageFileRepository
 from .file_repository_caching_decorator import FileRepositoryCachingDecorator
 from .file_repository_locking_decorator import FileRepositoryLockingDecorator
 from .file_repository_logging_decorator import FileRepositoryLoggingDecorator
-
-from .agent_plugin_repository_logging_decorator import AgentPluginRepositoryLoggingDecorator
-from .agent_plugin_repository_caching_decorator import AgentPluginRepositoryCachingDecorator
 
 from .file_simulation_repository import FileSimulationRepository
 from .mongo_credentials_repository import MongoCredentialsRepository
@@ -25,7 +21,6 @@ from .mongo_machine_repository import MongoMachineRepository
 from .mongo_agent_repository import MongoAgentRepository
 from .mongo_node_repository import MongoNodeRepository
 from .mongo_agent_event_repository import MongoAgentEventRepository
-from .file_agent_plugin_repository import FileAgentPluginRepository
 
 from .utils import initialize_machine_repository
 from .agent_machine_facade import AgentMachineFacade
