+++ monkey/monkey_island/cc/services/agent_plugin_service/agent_plugin_repository_caching_decorator.py
@@ -1,35 +1,32 @@
-import logging
+from functools import lru_cache
 from typing import Any, Dict
 
 from common import OperatingSystem
 from common.agent_plugins import AgentPlugin, AgentPluginManifest, AgentPluginType
 
-from . import IAgentPluginRepository
-
-logger = logging.getLogger(__name__)
+from .i_agent_plugin_repository import IAgentPluginRepository
 
 
-class AgentPluginRepositoryLoggingDecorator(IAgentPluginRepository):
+class AgentPluginRepositoryCachingDecorator(IAgentPluginRepository):
     """
-    An IAgentPluginRepository decorator that provides debug logging for other
-    IAgentPluginRepositories
+    An IAgentPluginRepository decorator that provides caching for other IAgentPluginRepositories
     """
 
     def __init__(self, agent_plugin_repository: IAgentPluginRepository):
         self._agent_plugin_repository = agent_plugin_repository
 
+    @lru_cache()
     def get_plugin(
         self, host_operating_system: OperatingSystem, plugin_type: AgentPluginType, name: str
     ) -> AgentPlugin:
-        logger.debug(f"Retrieving plugin {name} of type {plugin_type}")
         return self._agent_plugin_repository.get_plugin(host_operating_system, plugin_type, name)
 
+    @lru_cache()
     def get_all_plugin_configuration_schemas(
         self,
     ) -> Dict[AgentPluginType, Dict[str, Dict[str, Any]]]:
-        logger.debug("Retrieving plugin configuration schemas")
         return self._agent_plugin_repository.get_all_plugin_configuration_schemas()
 
+    @lru_cache()
     def get_all_plugin_manifests(self) -> Dict[AgentPluginType, Dict[str, AgentPluginManifest]]:
-        logger.debug("Retrieving plugin manifests")
         return self._agent_plugin_repository.get_all_plugin_manifests()
