+++ monkey/monkey_island/cc/services/agent_plugin_service/flask_resources/agent_plugins.py
@@ -7,8 +7,10 @@ from flask_security import auth_token_required, roles_accepted
 from common import OperatingSystem
 from common.agent_plugins import AgentPluginType
 from monkey_island.cc.flask_utils import AbstractResource
-from monkey_island.cc.repositories import IAgentPluginRepository, UnknownRecordError
+from monkey_island.cc.repositories import UnknownRecordError
 from monkey_island.cc.services.authentication_service import AccountRole
+
+from ..i_agent_plugin_repository import IAgentPluginRepository
 
 logger = logging.getLogger(__name__)
