+++ monkey/monkey_island/cc/services/agent_plugin_service/i_agent_plugin_repository.py
