+++ monkey/monkey_island/cc/services/agent_plugin_service/plugin_archive_parser.py
