+++ monkey/tests/unit_tests/monkey_island/cc/services/agent_configuration_service/test_agent_configuration_validation_decorator.py
@@ -2,7 +2,7 @@ import pytest
 from tests.monkey_island import InMemoryAgentConfigurationRepository, InMemoryAgentPluginRepository
 
 from common.base_models import MutableInfectionMonkeyBaseModel
-from monkey_island.cc.repositories import IAgentPluginRepository, RetrievalError
+from monkey_island.cc.repositories import RetrievalError
 from monkey_island.cc.services import PluginConfigurationValidationError
 from monkey_island.cc.services.agent_configuration_service.agent_configuration_schema_compiler import (  # noqa: E501
     AgentConfigurationSchemaCompiler,
@@ -31,7 +31,7 @@ def in_memory_agent_configuration_repository() -> IAgentConfigurationRepository:
 
 
 @pytest.fixture
-def in_memory_agent_plugin_repository() -> IAgentPluginRepository:
+def in_memory_agent_plugin_repository() -> InMemoryAgentPluginRepository:
     return InMemoryAgentPluginRepository()
