+++ monkey/tests/unit_tests/monkey_island/cc/services/agent_plugin_service/flask_resources/test_agent_plugins.py
@@ -10,8 +10,11 @@ from tests.unit_tests.monkey_island.cc.fake_agent_plugin_data import (
 )
 from tests.unit_tests.monkey_island.conftest import get_url_for_resource
 
-from monkey_island.cc.repositories import IAgentPluginRepository, RetrievalError
+from monkey_island.cc.repositories import RetrievalError
 from monkey_island.cc.services.agent_plugin_service.flask_resources import AgentPlugins
+from monkey_island.cc.services.agent_plugin_service.i_agent_plugin_repository import (
+    IAgentPluginRepository,
+)
 
 FAKE_PLUGIN_NAME = "plugin_abc"
